pull out properties viewer
